Move tests from tests/virtool_workflow -> tests/
