+++ tests/analysis/test_analysis_fixture.py
