+++ tests/analysis/test_indexes/test_write_isolate_fasta.txt
