+++ tests/api/mocks/mock_cache_routes.py
