+++ tests/api/mocks/mock_index_routes.py
