+++ tests/execution/test_fixture_hooks.py
